spark-238:remove unused code
